requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
ec2-metadata==2.11.0
Flask==2.2.3
idna==3.4
importlib-metadata==6.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
